Making changes to accomodate Mac usage.
Added build and upload instructions for the Mac environment.
Made some improvments to the Makefile such as:
 - Relative directory structure
 - Overloadable variables for more flexible usage
 - Some example alternative values added in comments
 - Added test node

Made some organizational updates to the README from styles I have seen, along with some consolidation for convenient brevity, including merging the two Makefiles.
Added some information to the README.  Tried to find a balance in verbosity and detail.